Splits code to seperate modules & add documention
