+++ tools/kcli/kcli/conf.py
@@ -1,15 +1,16 @@
 import ConfigParser
 import os
+import time
 
 from kcli import exceptions
-
-from kcli.exceptions import IRFileNotFoundException
 
 ENV_VAR_NAME = "KCLI_CONFIG"
 KCLI_CONF_FILE = 'kcli.cfg'
 CWD_PATH = os.path.join(os.getcwd(), KCLI_CONF_FILE)
 USER_PATH = os.path.expanduser('~/.' + KCLI_CONF_FILE)
 SYSTEM_PATH = os.path.join('/etc/khaleesi', KCLI_CONF_FILE)
+YAML_EXT = ".yml"
+TMP_OUTPUT_FILE = 'kcli_settings_' + str(time.time()) + YAML_EXT
 
 
 def load_config_file():
